refactor the moveKeyPoint function

+++ js/index.js
@@ -1,24 +1,27 @@
 var $ = require('jquery')
 import start from './helpFn/init'
 require('jquery-ui/droppable')
-import{ setUser } from './helpFn/setUser'
+import checkUser from './helpFn/checkAuthor'
+import {setUser} from './helpFn/setUser'
 
-var user;
 $(document).ready(function(){
   start()
   $('#bin').droppable({
   	drop: function( event, ui ) {
-  	 let id= $(ui.draggable).attr('id')
-  	 $.ajax({
-  	 	url:'/posts/'+id,
-  	 	method:'DELETE'
-  	 })
-  	 $(ui.draggable).remove()
+    if(checkUser(ui.draggable)){
+    	 let id= $(ui.draggable).attr('id')
+    	 $.ajax({
+    	 	url:'/posts/'+id,
+    	 	method:'DELETE'
+    	 })
+    	 $(ui.draggable).remove()
+     }
   	}
   })
 
   $('#user input').change(function(){
-    setUser($(this).val());
+
+    setUser( $(this).val() );
   })
 });
